setup-mssql/python/requirements.txt
pandas
pyodbc
ipykernel
python-dotenv
sqlalchemy